+++ templates/styled-component/template/src/documentation.js
@@ -1,0 +1,16 @@
+/**
+ * Options for {{classifiedName}}
+ * @typedef {Object} {{classifiedName}}Options
+ */
+
+/**
+ * {{description}}
+ * @class {{classifiedName}}
+ * @classdesc
+ * @property { {{~classifiedName~}}Options} options
+ * @example
+ * {% import {{classifiedName}} from '{{name}}' %}
+ *   ...
+ *   <{{classifiedName}}/>
+ *   ...
+ */